Made suds work with operations taking choice parameters.
Based on a patch by michaelgruenewald & bennetb01 attached to ticket #342 on the
original suds project issue tracker ('http://fedorahosted.org/suds/ticket/342').
Comments listed related to that ticket seem to indicate that there may be
additional problems with this patch but so far we have not encountered any.

Added a test to make sure that such choice parameters are correctly listed in
the corresponding suds method object.