add tests for client fetching documents using a registered transport
If an external document is not cached or found in the client's registered
document store, it should be fetched using the registered transport. This holds
for directly used & indirectly imported WSDL schemas, as well as imported or
included XSD schemas.

Related existing tests refactored a bit to reduce code duplication:
  * extracted shared test data
  * test_WSDL_not_found_in_cache_or_store_should_be_transported() renamed to
    test_WSDL_transport() and moved into the TestTransportUsage test class